Added dig function and fixed some things

+++ include/GameObject.h
@@ -13,19 +13,19 @@ class GameObject {
 
         //ritorna la coordinata x dell'oggetto di gioco
         //returns the game object's x coordinate
-        int getX(){ return x; }
+        float getX(){ return x; }
 
         //ritorna la coordinata y dell'oggetto di gioco
         //returns the game object's y coordinate
-        int getY(){ return y; }
+        float getY(){ return y; }
 
         //setta il valore della coordinata x dell'oggetto di gioco
         //sets the value of the x coordinate of the game object
-        int setX(int _x){x = _x;}
+        float setX(float _x){ x = _x; }
 
         //setta il valore della coordinata y dell'oggetto di gioco
         //sets the value of the y coordinate of the game object
-        int setY(int _y){y = _y;}
+        float setY(float _y){ y = _y; }
 
         //ritorna lo stato di visibilità dell'oggetto di gioco
         //returns the game object's visibility state
